fix: change user_id from Integer to BigInteger to handle large google_id
- Modify User.id and Fiche.user_id from Integer to BigInteger in models.py
- Add migrate_to_bigint() function to ALTER TABLE columns at startup
- Add sqlalchemy.text import for raw SQL execution
- This fixes 'value out of range for type integer' errors with google IDs
- JWT already uses user.id (integer) instead of google_id (string)

+++ backend/models.py
@@ -15,7 +15,7 @@ class User(db.Model):
     """
     __tablename__ = 'users'
 
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    id = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
     google_id = db.Column(db.String(255), unique=True, nullable=False)  # Google sub
     email = db.Column(db.String(255), unique=True, nullable=False)
     name = db.Column(db.String(255), nullable=False)
@@ -45,7 +45,7 @@ class Fiche(db.Model):
     __tablename__ = 'fiches'
 
     id = db.Column(db.String(255), primary_key=True, default=lambda: str(uuid.uuid4()))  # UUID or Google location ID
-    user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('users.id'), nullable=False)
     nom = db.Column(db.String(255), nullable=False)
     categorie = db.Column(db.String(255), nullable=True)
     adresse = db.Column(db.Text, nullable=True)
